+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* **BREAKING** Document Component: Meer dan 5 heading levels ondersteunen ([#3897](https://github.com/dso-toolkit/dso-toolkit/issues/3897))
+
 ### Changed
 * Document Component: `markFn` markeert onterecht tekst in tooltip ([#3865](https://github.com/dso-toolkit/dso-toolkit/issues/3865))